# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zhengji Wellness × Kiri Bar Referral Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: partnerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead (customer signup)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "409":
          description: Duplicate mobile or membership ID
...
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/stage:
    patch:
      operationId: updateLeadStage
      tags: [leads]
      summary: Update lead pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/verify-payment:
    post:
      operationId: verifyPayment
      tags: [leads]
      summary: Verify payment and trigger commission calculation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerification"
      responses:
        "200":
          description: Payment verified, commission calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/whatsapp-message:
    get:
      operationId: getLeadWhatsappMessage
      tags: [leads]
      summary: Get WhatsApp-ready confirmation messages for lead (both languages)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: WhatsApp messages in both languages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappMessagePair"
...
components:
  schemas:
...
    LeadInput:
      type: object
      required: [name, mobile, referralCode, consentGiven, selectedOffer]
      properties:
        name:
          type: string
        nameZh:
          type: string
        mobile:
          type: string
        whatsapp:
          type: string
        kirimembershipId:
          type: string
        referralCode:
          type: string
        consentGiven:
          type: boolean
        selectedOffer:
          type: string
        appointmentIntent:
          type: string
        preferredDate:
          type: ["string", "null"]
        notes:
          type: string
        lang:
          type: string
          enum: [en, zh]
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        whatsapp:
          type: string
        notes:
          type: string
        appointmentDate:
          type: ["string", "null"]
        auditNote:
          type: string
    StageUpdate:
      type: object
      required: [stage]
      properties:
        stage:
          type: string
          enum: [new_lead, appointment_booked, arrived, free_consultation_only, first_paid_treatment, package_purchased, invalid_cancelled]
        auditNote:
          type: string
    PaymentVerification:
      type: object
      required: [netSaleAmount, paymentType]
      properties:
        netSaleAmount:
          type: number
        paymentType:
          type: string
          enum: [first_paid_treatment, package_purchased]
        proofUrl:
          type: ["string", "null"]
        auditNote:
          type: string
        commissionType:
          type: string
          enum: [flat_rm30, package_percent]
        commissionRate:
          type: ["number", "null"]
    Lead:
      type: object
      required: [id, name, mobile, stage, referralCode, partnerId, consentGiven, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameZh:
          type: ["string", "null"]
        mobile:
          type: string
        whatsapp:
          type: ["string", "null"]
        kirimembershipId:
          type: ["string", "null"]
        referralCode:
          type: string
        partnerId:
          type: integer
        partnerName:
          type: string
        stage:
          type: string
          enum: [new_lead, appointment_booked, arrived, free_consultation_only, first_paid_treatment, package_purchased, invalid_cancelled]
        selectedOffer:
          type: string
        appointmentIntent:
          type: ["string", "null"]
        appointmentDate:
          type: ["string", "null"]
        netSaleAmount:
          type: ["number", "null"]
        proofUrl:
          type: ["string", "null"]
        consentGiven:
          type: boolean
        notes:
          type: ["string", "null"]
        lang:
          type: string
        commissionId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadList:
      type: object
      required: [leads, total, page, limit]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    WhatsappMessagePair:
      type: object
      required: [messageEn, messageZh]
      properties:
        messageEn:
          type: string
        messageZh:
          type: string
...